Migrate OSGi integration tests from Pax Exam to bnd
