+++ tests/common/src/main/resources/org/jupnp/osgi/test/data/initial-data.properties
