+++ tests/common/src/main/resources/org/jupnp/osgi/test/data/test-data-factory.properties
