+++ tests/devices/simple/bnd.bnd
@@ -1,0 +1,3 @@
+-include: ../../../bnd.bnd
+
+-includeresource: jupnp-osgi-tests-common-*.jar;lib:=true
