+++ tests/devices/simple/src/main/resources/org/jupnp/osgi/upnp/test/device/simple/icons/upnp-icon-large.png
