+++ tests/pom.xml
@@ -24,6 +24,8 @@
   <properties>
     <!-- This module (and its sub-modules) doesn't have a manual -->
     <lemma.manual>false</lemma.manual>
+    <maven.deploy.skip>true</maven.deploy.skip>
+    <maven.javadoc.skip>true</maven.javadoc.skip>
   </properties>
 
   <dependencies>
@@ -40,16 +42,5 @@
       <artifactId>osgi.cmpn</artifactId>
     </dependency>
   </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <artifactId>maven-deploy-plugin</artifactId>
-        <configuration>
-          <skip>true</skip>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
 
 </project>
